feat(wizard): wire live ChatBackendAdapter for LLM path (Task 5)
- Relocate ChatBackendAdapter from cmd/skill_generate.rs into
  conversations/backend/adapter.rs as a shared pub struct with
  build_chat_adapter() constructor
- Update skill_generate.rs to import from the shared module (deletes
  ~50 lines of local duplication)
- Wire build_chat_adapter(&mur_home, None, "agent.wizard") into
  wizard::run, replacing the Task-5 stub; falls back gracefully to
  deterministic stubs when no model is configured
